+++ src/Scenario.cpp
@@ -1,6 +1,6 @@
 #include "../headers/Scenario.h"
 
-Scenario::Scenario(float probability, vector<int> cases_per_block)
+Scenario::Scenario(float probability, vector<float> cases_per_block)
 {
     this->probability = probability;
     this->cases_per_block = cases_per_block;
